move jsonify_pg_structure
